refactor(AllDetailsTable): move component to its own directory
