+++ src/components/parameters/invoices/AllDetailsTable/AllDetailsTable.tsx
