+++ src/pages/transacciones/factura/detail/index.tsx
@@ -6,7 +6,7 @@ import PageTitle from '~components/titles/PageTitle'
 import { useGetOneInvoiceQuery } from '~redux/api/bca-backend/transacciones/invoiceSlice'
 import InvoiceForm from '~/components/forms/Invoice/Invoice'
 import EditToolbar from '~components/table/headers/toolbar'
-import AllDetailsTable from '~components/parameters/invoices/AllDetailsTable'
+import AllDetailsTable from '~/components/parameters/invoices/AllDetailsTable/AllDetailsTable'
 import { useGetAllInvoiceDetailsQuery } from '~redux/api/bca-backend/transacciones/invoiceDetailsSlice'
 import InvoiceDetailsDrawer from '~/components/drawers/Transactions/InvoiceDetailsDrawer/InvoiceDetailsDrawer'
